+++ data/src/main/java/com/anytypeio/anytype/data/auth/repo/block/BlockDataStore.kt
@@ -83,6 +83,7 @@ interface BlockDataStore {
     ): Pair<Id, Payload>
 
     suspend fun addRelationToDataView(ctx: Id, dv: Id, relation: Id): Payload
+    suspend fun deleteRelationFromDataView(ctx: Id, dv: Id, relation: Id): Payload
 
     suspend fun updateDataViewViewer(
         context: Id,