+++ domain/src/main/java/com/anytypeio/anytype/domain/block/repo/BlockRepository.kt
@@ -136,6 +136,7 @@ interface BlockRepository {
     ): Pair<Id, Payload>
 
     suspend fun addRelationToDataView(ctx: Id, dv: Id, relation: Id): Payload
+    suspend fun deleteRelationFromDataView(ctx: Id, dv: Id, relation: Id): Payload
 
     suspend fun updateDataViewViewer(
         context: Id,